backend/requirements.txt
# Backend API Requirements
# Python 3.10+

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# Email
sendgrid==6.11.0

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Generation
reportlab==4.0.9

# Testing (optional)
pytest==7.4.4
httpx==0.26.0
